fix: 지도 인스턴스 생성 타이밍 관련 버그 수정 (#98)
- `react-kakao-maps-sdk` 로드(loading)와 맵 인스턴스 생성(onCreate) 시점이 달라,
- `ref`의 `current` 속성이 `undefined`로 출력되는 이슈 해결
